doesnt return empty objects

+++ ai_rest_api.py
@@ -5,7 +5,7 @@ import sys
 from flask import Flask, request
 from collections import defaultdict
 from BayesNetwork import MultiNetwork
-from Translator import Translator, Flattener
+from Translator import Translator, Flattener, copy_attribute
 from gevent.pywsgi import WSGIServer
 from config import repo_dir, log_dir, path_pid
 import logging
@@ -13,12 +13,16 @@ import logging
 
 app = Flask(__name__)
 
-applogger = app.logger
-file_handler = logging.FileHandler(log_dir + "/app.log")
-file_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
-file_handler.setLevel(logging.DEBUG)
-applogger.setLevel(logging.DEBUG)
-applogger.addHandler(file_handler)
+
+logging.basicConfig(
+    filename=log_dir + "/app.log",
+    level=logging.DEBUG,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+
+# file_handler = logging.FileHandler(log_dir + "/app.log")
+# file_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
+# file_handler.setLevel(logging.DEBUG)
 
 dict_sessions = {}
 translator = Translator()
@@ -29,12 +33,12 @@ def predict():
     content_type = request.headers.get('Content-Type')
     if content_type == 'application/json':
         json = request.json
+
         # id_session = json['id_session']
         id_session = 0
-
         location = json["location"]
         id_bp = json['plan']['businessPlan_id']
-
+        logging.info(f"received business plan with id {id_bp}")
         if id_session not in dict_sessions:
             dict_sessions[id_session] = MultiNetwork()
 
@@ -47,7 +51,8 @@ def predict():
 
         bp['location'] = location
         bp['plan']['businessPlan_id'] = id_bp
-
+        copy_attribute(json, bp)
+        logging.info(f"processing successfull of business plan with id {id_bp}")
         return bp
     else:
         return 'Content-Type not supported!'
@@ -55,8 +60,13 @@ def predict():
 
 if __name__ == "__main__":
 
+    if os.path.exists(path_pid):
+        print("pid file exists, daemon already running, if not - delete pid file")
+        sys.exit(1)
+
     with open(path_pid, "w") as conn:
         conn.write(str(os.getpid()))
+
     parser = argparse.ArgumentParser()
     parser.add_argument('--ip', type=str, default="localhost")
     parser.add_argument('--port', type=int, default="2222")
